# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Management API",
        "description": "API documentation for the PMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API"
        }
    ],
    "paths": {
...
        "/labels": {
            "get": {
                "tags": [
                    "Labels"
                ],
                "summary": "List labels in workspace",
                "description": "Returns a list of all labels in the current workspace.",
                "operationId": "listLabels",
                "responses": {
                    "200": {
                        "description": "Labels retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Labels"
                ],
                "summary": "Create a new label",
                "description": "Creates a new label in the current workspace.",
                "operationId": "createLabel",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the label",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color for the label",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the label",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/labels/{label}": {
            "get": {
                "tags": [
                    "Labels"
                ],
                "summary": "Get label details",
                "description": "Returns details of a specific label.",
                "operationId": "getLabel",
                "parameters": [
                    {
                        "name": "label",
                        "in": "path",
                        "description": "The ID of the label",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Label retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Labels"
                ],
                "summary": "Update a label",
                "description": "Updates an existing label.",
                "operationId": "updateLabel",
                "parameters": [
                    {
                        "name": "label",
                        "in": "path",
                        "description": "The ID of the label",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the label",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color for the label",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the label",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Labels"
                ],
                "summary": "Delete a label",
                "description": "Deletes an existing label.",
                "operationId": "deleteLabel",
                "parameters": [
                    {
                        "name": "label",
                        "in": "path",
                        "description": "The ID of the label",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Label deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}